Fix: move graphviz-dot tests out to conditional module
